Chris' patch for MatrixView plus some other tiny fiddling.
"This patch:

 * sorts out the problems with matrix editing of segments with
   non-zero start times
 * improves the behaviour of the right-hand end of the matrix
   when you attempt to drag notes off the end (actually just
   by preventing you from doing so)
 * fixes undo and redo for note insertion
 * fixes the inserter for drag-left (i.e. end time before start
   time)
 * centralises the settings for matrix drag resolution in a
   single SnapGrid object in matrixview (calling setSnapTime
   on MatrixView's m_snapGrid should change the snap time for
   all tools)
 * removes a few compiler warnings
 * adds a bit more detail to the faq
 * adds somewhat more text to the docbook help"


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@2707 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/editview.cpp
@@ -353,14 +353,14 @@ void EditView::slotScrollHorizSmallSteps(int hpos)
 	// returning to zero
         hbar->setValue(0);
 
-    } else if ((diff = hpos - (scrollView->contentsX() + 
-			       scrollView->visibleWidth() * 0.9)) > 0) {
+    } else if ((diff = int(hpos - (scrollView->contentsX() + 
+				   scrollView->visibleWidth() * 0.9))) > 0) {
 
 	// moving off the right hand side of the view   
 	hbar->setValue(hbar->value() + diff);
 
-    } else if ((diff = hpos - (scrollView->contentsX() +
-			       scrollView->visibleWidth() * 0.1)) < 0) {
+    } else if ((diff = int(hpos - (scrollView->contentsX() +
+				   scrollView->visibleWidth() * 0.1))) < 0) {
 
 	// moving off the left
 	hbar->setValue(hbar->value() + diff);
